server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==0.21.0
pydantic!=1.8,!=1.8.1,!=2.0.0,!=2.0.1,!=2.1.0,<3.0.0
supabase==2.9.0
bcrypt
resend
PyJWT